change: put batch metadata update in side panel too
